refactor: update UsersRepository import paths to reflect membership module structure
